Create SpatialVxReferencesList.md
List from NCAR's project.ral.ucar.edu/icp web page, which I expect will not be very stable going forward.  Also, now we can add to it.  I did not put the links to these papers as it is too time consuming.  Nevertheless, it's not difficult to find the sources these days.